requirements.txt
pytest>=7.1.2

allure-pytest>=2.9.45

selene>=2.0.b4

h11~=0.14.0

pip~=22.3.1

attrs~=23.1.0

wheel~=0.38.4

setuptools~=65.5.1

future~=0.18.3

requests~=2.31.0

colorama~=0.4.6

trio~=0.22.2

outcome~=1.3.0.post0

sniffio~=1.3.0

sortedcontainers~=2.4.0

cffi~=1.16.0

idna~=3.4

selenium~=4.14.0

pluggy~=1.3.0

iniconfig~=2.0.0

certifi~=2023.7.22

urllib3~=2.0.7

pycparser~=2.21

packaging~=23.2

wsproto~=1.2.0

tqdm~=4.66.1

python-dotenv~=1.0.0

pyvirtualdisplay>=3.0